change type to int

+++ src/main/java/phonerate/Silver.java
@@ -2,13 +2,13 @@ package phonerate;
 
 public class Silver {
 	private double silverrateperadditionalminute = 0.54;
-	private double silverincludedminutes = 500;
+	private int silverincludedminutes = 500;
 	private double silverrateperadditionalline = 21.50;
 	private double silverplanrate = 29.95;
 	public double rateperadditionalminute() {
 		return silverrateperadditionalminute;
 	}
-	public double includedminutes() {
+	public int includedminutes() {
 		return silverincludedminutes;
 	}
 	public double rateperadditionalline() {
